feat(partnership) : 파트너십 이미지 API 메서드 구현 및 AdminController로 분리
