+++ src/main/java/com/terbuck/terbuck_be/domain/partnership/controller/PartnershipController.java
@@ -1,4 +1,4 @@
-package com.terbuck.terbuck_be.domain.partnership;
+package com.terbuck.terbuck_be.domain.partnership.controller;
 
 import com.terbuck.terbuck_be.common.dto.SuccessMessage;
 import com.terbuck.terbuck_be.common.dto.SuccessStatusResponse;
@@ -12,7 +12,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 @RestController
 @RequiredArgsConstructor
@@ -20,7 +19,6 @@ import org.springframework.web.multipart.MultipartFile;
 public class PartnershipController {
 
     private final PartnershipService partnershipService;
-    private final CsvPartnershipImporter csvPartnershipImporter;
 
     @GetMapping("/home")
     public ResponseEntity<SuccessStatusResponse<PartnershipListResponse<HomePartnershipDto>>> getHomeShop(
@@ -41,14 +39,5 @@ public class PartnershipController {
                 .body(SuccessStatusResponse.of(SuccessMessage.ID_PARTNERSHIP_GET_SUCCESS, partnershipResponse));
     }
 
-    @PostMapping("/upload")
-    public ResponseEntity<String> uploadCsv(@RequestParam("file") MultipartFile file, @RequestParam("university") University university) {
-        try {
-            csvPartnershipImporter.importFromCsv(file, university);
-            return ResponseEntity.ok("업로드 및 저장 성공!");
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body("업로드 실패: " + e.getMessage());
-        }
-    }
+
 }
