+++ src/main/java/com/terbuck/terbuck_be/domain/partnership/dto/UpdateImageRequest.java
@@ -1,0 +1,10 @@
+package com.terbuck.terbuck_be.domain.partnership.dto;
+
+import com.terbuck.terbuck_be.common.enums.University;
+import lombok.Data;
+
+@Data
+public class UpdateImageRequest {
+
+    private University university;
+}
